requirements.txt
Django==3.2.6
django-environ==0.4.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
django-jazzmin
djoser==2.1.0
Pillow==8.3.1
psycopg2-binary
uwsgi
wheel
qrcode==7.2
